# src/swagger.yml
swagger: "2.0"
info:
  description: "This is ClouDreamer's Multicloud Logging & Monitoring system API document."
  version: "1.1.0"
  title: "LM API"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "ejb0329@gmail.com"
basePath: "/"
tags:
- name: "logging"
  description: "API about logging system"
- name: "monitoring"
  description: "API about monitoring system"
- name: "auth"
  description: "API about authentication"
schemes:
- "https"
- "http"
paths:
  /sync:
    get:
      tags:
      - "logging"
      summary: "Synchronize and save logs"
      description: "Compare the date of the last row stored in the database and the log and save it."
      consumes:
      - "application/json"
      produces:
      - "string"
      responses:
        "200":
          description: "Successful operation"
          schema:
            type: "string"
            example: "ok"
        "400":
          description: "Bad ElasticSerach api request"
        "500":
          description: "Unable to parse time"
  /{component}/log:
    get:
      tags:
      - "logging"
      summary: "Get logs of each component"
      description: "Returns a list of logs"
      produces:
      - "application/json"
      parameters:
      - name: "component"
        in: "path"
        description: "Name of component to return"
        required: true
        type: "string"
      - name: "_start"
        in: "query"
        description: "Start Index"
        required: true
        type: "integer"
      - name: "_end"
        in: "query"
        description: "End index"
        required: true
        type: "integer"
      responses:
        "200":
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Log"
    delete:
      tags:
      - "logging"
      summary: "Delete logs of each component"
      description: "Delete logs"
      produces:
      - "application/json"
      parameters:
      - name: "component"
        in: "path"
        description: "Name of component to delete"
        required: true
        type: "string"
      responses:
        "200":
          description: "Successful operation"
          schema:
            type: "string"
            example: "clear success"
        "400":
          description: "Unable to delete ElasticSearch index"
  /check:
    get:
      tags:
      - "logging"
      summary: "check error component"
      description: "Check the component where the error occurred(true: occured, false: not occured)"
      produces:
      - "application/json"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ErrorComponent"
  /token:
    post:
      tags:
      - "auth"
      summary: "Get OpenStack scoped token"
      description: "Returns scoped token for entered user and project."
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "User login inforamtion of openstack"
        required: true
        schema:
          $ref: "#/definitions/LoginOpenstack"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Token"
        "400":
          description: "openstack api request error"
  /instances:
    get:
      tags:
      - "monitoring"
      summary: "Get OpenStack instance list"
      description: "Returns OpenStack instance list(CloudStack part is not yet)"
      produces:
      - "application/json"
      parameters:
      - in: "header"
        name: "token"
        type: "string"
        required: true
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Instances"
  /statistics:
    post:
      tags:
      - "monitoring"
      summary: "Get metrics"
      description: "Returns statistics from OpenStack and CloudStack"
      produces:
      - "application/json"
      parameters:
      - in: "header"
        name: "token"
        type: "string"
        required: true
      - in: "body"
        name: "body"
        description: "User inforamtion(key) of cloudstack"
        required: true
        schema:
          $ref: "#/definitions/CloudstackReq"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Statistics"
        "400":
          description: "openstack and cloudstack api request error"
        
definitions:
  Log:
    type: "array"
    items:
      $ref: "#/definitions/LogDetail"
  LogDetail:
    type: "object"
    properties:
      id:
        type: "integer"
      created_on:
        type: "string"
      component:
        type: "string"
      level:
        type: "string"
      message:
        type: "string"
  ErrorComponent:
    type: "object"
    properties:
      id:
        type: "integer"
      nova:
        type: "boolean"
        default: false
      heat:
        type: "boolean"
        default: false
      cinder:
        type: "boolean"
        default: false
      neutron:
        type: "boolean"
        default: false
      keystone:
        type: "boolean"
        default: false
      swift:
        type: "boolean"
        default: false
      agent:
        type: "boolean"
        default: false
      management:
        type: "boolean"
        default: false
  LoginOpenstack:
    type: "object"
    properties:
      name:
        type: "string"
      password:
        type: "string"
      project_id:
        type: "string"
  Token:
    type: "object"
    properties:
      token:
        type: "string"
  Instances:
    type: "object"
    properties:
      servers:
        type: "array"
        items: 
          $ref: "#/definitions/Server"
      servers_links:
        type: "array"
        items:
          $ref: "#/definitions/Link"
  Server:
    type: "object"
    properties:
      id:
        type: "string"
      name:
        type: "string"
      links:
        type: "array"
        items: 
          $ref: "#/definitions/Link"
  Link:
    type: "object"
    properties:
      href:
        type: "string"
      rel:
        type: "string"
  CloudstackReq:
    type: "object"
    properties:
      requests:
        type: "object"
        properties:
          apikey:
            type: "string"
          response:
            type: "string"
            default: "json"
          command:
            type: "string"
            default: "listCapacity"
      secretKey:
        type: "string"
  Statistics:
    type: "object"
    properties:
      id:
        type: "integer"
      openstack_metrics:
        type: "object"
        properties:
          vcpu:
            type: "integer"
          vcpu_used:
            type: "integer"
          memory:
            type: "integer"
          memory_used:
            type: "integer"
          storage:
            type: "integer"
          storage_used:
            type: "integer"
        example:
          vcpu: 4
          vcpu_used: 0
          memory: 9965
          memory_used: 512
          storage: 48
          storage_used: 0
      cloudstack_metrics:
        type: "object"
        properties:
          vcpu:
            type: "integer"
          vcpu_used:
            type: "integer"
          memory:
            type: "integer"
          memory_used:
            type: "integer"
          storage:
            type: "integer"
          storage_used:
            type: "integer"
        example:
          vcpu: 16
          vcpu_used: 3
          memory: 32354942976
          memory_used: 1879048192
          storage: 53660876800
          storage_used: 26289897472

externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"